Fixed two issues
First issue is that the os uses the scripts location as the execution folder, I changed so that it will change the directory to the directory that is given as an argument.
Second issue is that files with dots in their names get a wrong file extension( i.g. "songSong 28.7.1888.mp3" will be recognized as 28 or 7), I changed the script to take the substring that is after the last dot and it fixed it.